further improved multi threading class functionality with better security and methods, added functionality to save tile map

+++ building.cpp
@@ -83,7 +83,7 @@ building::building(real32 PosX, real32 PosY, real32 PosZ, real32 SizeX, real32 S
 	Texture = new texture(L"data/2x3_texture-2.png", 256, 256, 1, 1);
 
 	VertexBuffer = new vertex_buffer(Vertices.data(), sizeof(vertex), (uint32)Vertices.size(), STATIC);
-	VertexBuffer->AddIndexBuffer(Indices.data(), sizeof(uint16), (uint32)Indices.size());
+	VertexBuffer->AddIndexBuffer(Indices.data(), sizeof(uint32), (uint32)Indices.size());
 
 	ConstantBuffer = new constant_buffer(&VS_Input, sizeof(cb));
 }
